requirements.txt
streamlit
streamlit-tags
requests
fastapi
"uvicorn[standard]"
langchain-openai 
langchain 
playwright 
beautifulsoup4
cohere
python-dotenv